add InvalidCurrencyPair handle
