+++ api/app.py
@@ -1,18 +1,8 @@
-import datetime
-from flask import Flask, make_response, Response, jsonify, request, Blueprint
-from flask_restx import Resource, Api, fields
+from flask import make_response, Response, request
+from flask_restx import Resource
 from api.oracle_framework import OracleFramework
-from api.errors import errors
-
-app = Flask(__name__)
-app.register_blueprint(errors)
-api = Api(
-    app,
-    doc='/swagger',
-    version='1.0.0',
-    title='kylin-api',
-    description='The API to provide the price feeds for the currency pairs from different sources(coingecko, coinbase, kraken, cryptowatch, etc)',
-)
+from api.restx import api, app
+from api.errors import InvalidCurrencyPair
 
 oracle_framework = OracleFramework()
 
@@ -22,7 +12,10 @@ class PriceList(Resource):
     def get(self):
         currency_pairs = request.args['currency_pairs']
         prices = oracle_framework.get_prices(currency_pairs)
-        return make_response(prices, 200)
+        if oracle_framework.has_results(prices):
+            return make_response(prices, 200)
+        else:
+            raise InvalidCurrencyPair("We didn't find any results for your query. Ensure you are using valid currency symbols with format: '<FROM-CURRENCY>_<TO-CURRENCY>'. If passing multiple currency pairs, separate them with commas. Example: btc_usd,eth_gbp,kyl_jpy")
 
 @api.route('/health', endpoint='health')
 class HealthList(Resource):
